+++ aws-sdk-core/spec/aws/plugins/retry_errors_spec.rb
@@ -163,12 +163,20 @@ module Aws
 
         let(:credentials) { Credentials.new('akid', 'secret') }
 
+        let(:cache) { EndpointCache.new }
+
+        let(:api) { Seahorse::Model::Api.new }
+
         let(:config) {
           cfg = Seahorse::Client::Configuration.new
           cfg.add_option(:credentials, credentials)
+          cfg.add_option(:endpoint_cache, cache)
+          cfg.add_option(:api, api)
           RetryErrors.new.add_options(cfg)
           cfg.build!
         }
+
+        let(:operation) { Seahorse::Model::Operation.new }
 
         let(:resp) { Seahorse::Client::Response.new }
 
@@ -176,6 +184,8 @@ module Aws
 
         before(:each) do
           resp.context.config = config
+          operation.endpoint_discovery = {}
+          resp.context.operation = operation
           resp.context.http_response.status_code = 400
         end
 
@@ -183,6 +193,20 @@ module Aws
           allow(Kernel).to receive(:sleep)
           handler.handler = send_handler || block
           handler.call(resp.context)
+        end
+
+        it 'retries if endpoint discovery error is detected' do
+          config.api.endpoint_operation = :describe_endpoints
+          DescribeEndpointsRequest = Seahorse::Model::Shapes::StructureShape.new(
+            name:'DescribeEndpointsRequest')
+          config.api.add_operation(:describe_endpoints, Seahorse::Model::Operation.new.tap do |o|
+            o.endpoint_operation = true
+            o.input = Seahorse::Model::Shapes::ShapeRef.new(shape: DescribeEndpointsRequest)
+          end)
+          resp.error = Errors::EndpointDiscoveryError.new
+
+          handle { |context| resp }
+          expect(resp.context.retries).to eq(3)
         end
 
         it 'does not retry responses that have no error' do